Renamed monkey_island.cc.utils to monkey_island.cc.network_utils as it contains only networking related code
