+++ monkey/monkey_island/cc/models/monkey.py
@@ -8,7 +8,7 @@ import ring
 from monkey_island.cc.models.monkey_ttl import MonkeyTtl, create_monkey_ttl_document
 from monkey_island.cc.consts import DEFAULT_MONKEY_TTL_EXPIRY_DURATION_IN_SECONDS
 from monkey_island.cc.models.command_control_channel import CommandControlChannel
-from monkey_island.cc.utils import local_ip_addresses
+from monkey_island.cc.network_utils import local_ip_addresses
 from common.cloud import environment_names
 
 MAX_MONKEYS_AMOUNT_TO_CACHE = 100
