+++ monkey/monkey_island/cc/services/reporting/report.py
@@ -16,7 +16,7 @@ from monkey_island.cc.services.node import NodeService
 from monkey_island.cc.services.reporting.pth_report import PTHReportService
 from monkey_island.cc.services.reporting.report_exporter_manager import ReportExporterManager
 from monkey_island.cc.services.reporting.report_generation_synchronisation import safe_generate_regular_report
-from monkey_island.cc.utils import local_ip_addresses, get_subnets
+from monkey_island.cc.network_utils import local_ip_addresses, get_subnets
 
 __author__ = "itay.mizeretz"
